Seção 17: Palestrante e Rede Social - FrontEnd

+++ Back/src/ProEventos.API/Helpers/Util.cs
@@ -18,10 +18,11 @@ namespace ProEventos.API.Helpers
         }
 
         public void DeleteImage(string imageName, string destino){
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
-
-            if(System.IO.File.Exists(imagePath))
-                System.IO.File.Delete(imagePath);
+            if(!string.IsNullOrEmpty(imageName)){
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
+                if(System.IO.File.Exists(imagePath))
+                    System.IO.File.Delete(imagePath);
+            }
         }
 
         public async Task<string> SaveImage(IFormFile imageFile, string destino)
